fixed for compilation errors

+++ test/cl/FillTest/FillTest.cpp
@@ -2128,11 +2128,11 @@ TEST (simpleTest, basicDataBoltClDevVectAutoConvertCheck)
     ////////////////////////////////////////////////////
 
     std::fill(hv.begin(), hv.end(), static_cast< int >( dValue ) ); 
-    bolt::cl::fill(dv.begin(), dv.end(), dValue);
+    bolt::cl::fill(dv.begin(), dv.end(), static_cast< int >(dValue));
     cmpArrays(hv,dv);
 
     std::fill(hv.begin(), hv.end(), static_cast< int >( fValue ) ); 
-    bolt::cl::fill(dv.begin(), dv.end(), fValue);
+    bolt::cl::fill(dv.begin(), dv.end(), static_cast< int >(fValue));
     cmpArrays(hv,dv);
 
     bolt::cl::fill(ddv.begin(), ddv.end(),iValue);
@@ -2214,11 +2214,11 @@ TEST (AutomaticsimpleTest, basicDataBoltClDevVectAutoConvertCheck)
     ////////////////////////////////////////////////////
 
     std::fill(hv.begin(), hv.end(), static_cast< int >( dValue ) ); 
-    bolt::cl::fill(ctl, dv.begin(), dv.end(), dValue);
+    bolt::cl::fill(ctl, dv.begin(), dv.end(),  static_cast< int >( dValue));
     cmpArrays(hv,dv);
 
     std::fill(hv.begin(), hv.end(), static_cast< int >( fValue ) ); 
-    bolt::cl::fill(ctl, dv.begin(), dv.end(), fValue);
+    bolt::cl::fill(ctl, dv.begin(), dv.end(),  static_cast< int >( fValue));
     cmpArrays(hv,dv);
 
     bolt::cl::fill(ctl, ddv.begin(), ddv.end(),iValue);
@@ -2300,11 +2300,11 @@ TEST (SerialsimpleTest, basicDataBoltClDevVectAutoConvertCheck)
     ////////////////////////////////////////////////////
 
     std::fill(hv.begin(), hv.end(), static_cast< int >( dValue ) ); 
-    bolt::cl::fill(ctl, dv.begin(), dv.end(), dValue);
+    bolt::cl::fill(ctl, dv.begin(), dv.end(), static_cast< int >(dValue));
     cmpArrays(hv,dv);
 
     std::fill(hv.begin(), hv.end(), static_cast< int >( fValue ) ); 
-    bolt::cl::fill(ctl, dv.begin(), dv.end(), fValue);
+    bolt::cl::fill(ctl, dv.begin(), dv.end(), static_cast< int >(fValue));
     cmpArrays(hv,dv);
 
     bolt::cl::fill(ctl, ddv.begin(), ddv.end(),iValue);
@@ -2387,11 +2387,11 @@ TEST (MultiCoresimpleTest, basicDataBoltClDevVectAutoConvertCheck)
     ////////////////////////////////////////////////////
 
     std::fill(hv.begin(), hv.end(), static_cast< int >( dValue ) ); 
-    bolt::cl::fill(ctl, dv.begin(), dv.end(), dValue);
+    bolt::cl::fill(ctl, dv.begin(), dv.end(), static_cast< int >(dValue));
     cmpArrays(hv,dv);
 
     std::fill(hv.begin(), hv.end(), static_cast< int >( fValue ) ); 
-    bolt::cl::fill(ctl, dv.begin(), dv.end(), fValue);
+    bolt::cl::fill(ctl, dv.begin(), dv.end(),  static_cast< int >(fValue));
     cmpArrays(hv,dv);
 
     bolt::cl::fill(ctl, ddv.begin(), ddv.end(),iValue);
